requirements.txt
pinecone-client
openai
requests